Fix bug in CChildVisualizer::GetItems (#121)
This PR fixes a bug in CChildVisualizer::GetItems where the wrong variable was used to calculate the index. This is #120 

This also fixes a problem that I noticed where apparently the child process debugger now requires the `CommandLine` field in .ChildProcessDbgSettings files. So, this updates our files to include that.